- Fix for .mat output-format on 64-bit Linux (use uint32_t instead of long)
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@7958 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/simulation_result_mat.cpp
@@ -35,6 +35,7 @@
 #include "simulation_result_mat.h"
 #include <cstring>
 #include <cstdlib>
+#include <stdint.h>
 
 
 simulation_result_mat::simulation_result_mat(const char* filename, 
@@ -221,11 +222,11 @@ void simulation_result_mat::writeMatVer4MatrixHeader(const char *name,
                  bool is_text)
 {
   typedef struct {
-    long type;
-    long mrows;
-    long ncols;
-    long imagf;
-    long namelen;
+    uint32_t type;
+    uint32_t mrows;
+    uint32_t ncols;
+    uint32_t imagf;
+    uint32_t namelen;
   } MHeader_t;
   const int endian_test = 1;
   MHeader_t hdr;
